# Improve ansi buffer PrintableRuneWidth performance