--- a/ansi/buffer_test.go
+++ b/ansi/buffer_test.go
@@ -17,6 +17,19 @@ func TestBuffer_PrintableRuneWidth(t *testing.T) {
 	}
 }
 
+// go test -bench=BenchmarkBuffer_PrintableRuneWidth -benchmem -count=4
+func BenchmarkBuffer_PrintableRuneWidth(b *testing.B) {
+	var buf Buffer
+	buf.Buffer.WriteString("\x1B[38;2;249;38;114m(\x1B[0m\x1B[38;2;248;248;242mjust another test\x1B[38;2;249;38;114m)\x1B[0m")
+	b.RunParallel(func(pb *testing.PB) {
+		b.ReportAllocs()
+		b.ResetTimer()
+		for pb.Next() {
+			buf.PrintableRuneWidth()
+		}
+	})
+}
+
 // go test -bench=Benchmark_PrintableRuneWidth -benchmem -count=4
 func Benchmark_PrintableRuneWidth(b *testing.B) {
 	s := "\x1B[38;2;249;38;114mfoo"