@kiyonlin 2020-10-21
> Interesting approach!
> 
> I think we'll keep this on the backburner though, as it's not an order of magnitude faster. In such cases I think I prefer to keep the code simple(r). Zero allocations are obviously another argument for it, so I'll keep this PR in mind.

Makes sense.

We can see if it helps in package `wordwrap` because it depends on `ansi Buffer.PrintableRuneWidth`