@kiyonlin 2020-10-22
Improvement for wordwrap performance

OLD:
```hs
BenchmarkWordWrapString-8   	  801043	      1330 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  924964	      1347 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1317 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  990302	      1317 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  869684	      1355 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  933944	      1350 ns/op	     879 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  912145	      1353 ns/op	     876 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1336 ns/op	     873 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  768968	      1320 ns/op	     867 B/op	      11 allocs/op
BenchmarkWordWrapString-8   	  927973	      1358 ns/op	     879 B/op	      11 allocs/op
```

NEW:
```hs
BenchmarkWordWrapString-8   	 1000000	      1021 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1037 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1041 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1066 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1139982	      1031 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1060 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1062 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1066 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1094 ns/op	       0 B/op	       0 allocs/op
BenchmarkWordWrapString-8   	 1000000	      1042 ns/op	       0 B/op	       0 allocs/op
```

benchstat:
```hs
name              old time/op    new time/op    delta
WordWrapString-8    1.34µs ± 2%    1.05µs ± 4%   -21.39%  (p=0.000 n=10+10)

name              old alloc/op   new alloc/op   delta
WordWrapString-8      878B ± 1%        0B       -100.00%  (p=0.000 n=9+10)

name              old allocs/op  new allocs/op  delta
WordWrapString-8      11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
```